pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "Bed-Ready: Hospital Bed Reservation System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "uvicorn[standard]>=0.42.0",
    "beanie>=1.26.0",
    "motor>=3.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.43.0",
    "apscheduler>=3.10.0",
    "websockets>=12.0",
    "pydantic[email]>=2.12.5",
    "httpx>=0.28.1",
]
